% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVarEEclosestImage.R
\name{addVarEEclosestImage}
\alias{addVarEEclosestImage}
\title{Add variable from closest image in a Google Earth Engine Collection}
\usage{
addVarEEclosestImage(
  ee_feats,
  collection,
  reducer = NULL,
  maxdiff,
  bands = NULL,
  unmask = FALSE,
  monitor = TRUE
)
}
\arguments{
\item{ee_feats}{A feature collection we want to annotate with GEE data.
We need to upload an sf object with the features to GEE. This object must
also have character column with the dates that need to be matched against the
image collection dates. The format must be "yyyy-mm-dd" and the column must
be named "Date".}

\item{collection}{Either a character string with the name of the collection
we want to use or a GEE collection produced with \code{ee$ImageCollection()}.
See \href{https://developers.google.com/earth-engine/datasets/catalog}{GEE catalog}.}

\item{reducer}{A character string specifying the function apply when
extracting values for each feature. It is common to use "mean", "sum" or
"count". But there are many other, see 'ee.Reducer' under Client Libraries at
\url{https://developers.google.com/earth-engine/apidocs}.}

\item{maxdiff}{Maximum difference in days allowed for an image to be matched with
data.}

\item{bands}{Select specific bands from the image. Only one band at a time is
allowed for now.}

\item{unmask}{GEE masks missing values, which means they are not used for
computing means, counts, etc. Sometimes we might want to avoid this behaviour
and use 0 instead of NA. If so, set unmask to TRUE.}

\item{monitor}{Logical. If TRUE (default) monitoring messages produced
by \code{rgee} will displayed. If FALSE, only high-level messages will be displayed.}
}
\value{
A dataframe similar to \code{ee_feats} with variables added from the
\code{bands} selected from \code{collection}. Note that following \href{https://github.com/r-spatial/rgee}{rgee}
the name of the new variables will be the selected band (\code{bands} or else
all bands from \code{collection} followed by the spatial reducer \code{reducer}.
}
\description{
Each feature in a feature collection (i.e. row in asimple feature
collection) is matched with the image in an image collection that is closest
to it in time.
}
\examples{
\dontrun{
# Load the remote data asset
ee_points <- ee$FeatureCollection(pointAssetId)  # assetId must correspond to an asset in your GEE account

# Annotate points with TerraClimate dataset
points_tmmn <- addVarEEclosestImage(ee_feats = ee_points,
                                   collection = "IDAHO_EPSCOR/TERRACLIMATE",
                                   maxdiff = 15,
                                   bands = "tmmn")

# Annotate polygons with TerraClimate dataset
ee_polgs <- ee$FeatureCollection(polgsAssetId)  # assetId must correspond to an asset in your GEE account

features_tmmn <- addVarEEclosestImage(ee_feats = ee_polgs,
                                   collection = "IDAHO_EPSCOR/TERRACLIMATE",
                                   maxdiff = 15,
                                   reducer = "mean",
                                   bands = "tmmn")
}
}
