% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVarEEimage.R
\name{addVarEEimage}
\alias{addVarEEimage}
\title{Add variable from Google Earth Engine Image}
\usage{
addVarEEimage(
  ee_feats,
  image,
  reducer = NULL,
  bands = NULL,
  unmask = FALSE,
  monitor = TRUE,
  transfer = list(via = "drive", container = "rgee_backup")
)
}
\arguments{
\item{ee_feats}{A feature collection with the sites we want to annotate.
We should have uploaded an sf object with the sites to GEE, previously.
See \code{\link{uploadFeaturesToEE}}}

\item{image}{Either a character string with the name of the image we want to
use or an GEE image produced with \code{ee$Image()}. See
\href{https://developers.google.com/earth-engine/datasets/catalog}{GEE catalog}.}

\item{reducer}{Optional. A character string specifying the function apply when
extracting values for each sites. This is unnecessary when annotating points, but
you might have uploaded polygons instead. This reducer will compute a spatial
summary of the environmental layer within the polygon. It is common to use "mean", "sum" or
"count". But there are many other, see 'ee.Reducer' under Client Libraries at
\url{https://developers.google.com/earth-engine/apidocs}.}

\item{bands}{Select specific bands from the image. If not specified, sites
will be annotated with all bands (one column for each band).}

\item{unmask}{GEE masks missing values, which means they are not used for
computing means, counts, etc. Sometimes we might want to avoid this behaviour
and use 0 instead of NA. If so, set unmask to TRUE.}

\item{monitor}{Logical. If TRUE (default) monitoring messages produced
by \code{rgee} will displayed. If FALSE, only high-level messages will be displayed.}

\item{transfer}{A list with parameters used to communicate to and from GEE. It
must have to elements named \code{via} and \code{container}. See \code{\link[rgee]{ee_as_sf}}}
}
\value{
A dataframe similar to \code{ee_sites} with variables added from the
\code{bands} selected from \code{collection}. Note that following \href{https://github.com/r-spatial/rgee}{rgee}
the name of the new variables will be: the selected band (\code{bands} or else
all bands from \code{collection} followed by the spatial reducer \code{reducer}, if present.
}
\description{
Add variable from Google Earth Engine Image
}
\examples{
\dontrun{
# Load the remote data asset
ee_points <- ee$FeatureCollection(pointAssetId)  # assetId must correspond to an asset in your GEE account

# Annotate sites with the surface water occurrence image
sites_tmmn <- addVarEEcollection(ee_feats = ee_points,
                                 image = "JRC/GSW1_3/GlobalSurfaceWater",
                                 bands = "occurrence")

# We can also annotate polygons using the mean surface water occurrence within
# the polygon
ee_polgs <- ee$FeatureCollection(polgsAssetId)  # assetId must correspond to an asset in your GEE account

sites_tmmn <- addVarEEcollection(ee_feats = ee_polgs,
                                 image = "JRC/GSW1_3/GlobalSurfaceWater",
                                 reducer = "mean",
                                 bands = "occurrence")
}
}
